Move get_git_data to corresponding function in input file
